renamed form folders to be general. Testing form on router rn as homepage.
