+++ src/components/EventForm/EventForm.js
@@ -7,7 +7,7 @@ import Paper from '@material-ui/core/Paper';
 import Button from '@material-ui/core/Button';
 import Link from '@material-ui/core/Link';
 import Typography from '@material-ui/core/Typography';
-import MyFormBody from './MyFormBody';
+import FormBody from './FormBody';
 //TODO: import props and require the correct type
 
 
@@ -85,7 +85,7 @@ export default function MyForm() {
           </Typography>
           <br></br>
 
-          <MyFormBody/>
+          <FormBody/>
           <Button
             variant="contained"
             color="primary"
